Separate defaults for snippet and hardcore jquery version
