+++ src/browser/bundles/defaults.js
@@ -1,0 +1,4 @@
+import { version } from '../../defaults.js'
+
+export const cdnHost = 'cdn.rollbar.com';
+export const defaultRollbarJsUrl = `https://${cdnHost}/rollbarjs/refs/tags/v${version}/rollbar.min.js`;
